# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Leerr API
  version: "0.2"
  description: |
    Same-origin JSON API for the Leerr web app.

    Authentication: web sessions use an HttpOnly, SameSite=Strict cookie
    (`__Host-leerr` over HTTPS) and must send `X-CSRF-Token` on every
    non-GET request. Native sessions send `Authorization: Bearer <token>`
    and need no CSRF token. Each session kind only works through its own channel.

    Every error has the body `{"error": {"code": "...", "message": "..."}}`;
    messages are written for end users.
servers:
  - url: /
security:
  - cookie: []
  - bearer: []
...
paths:
...
  /api/v1/admin/users:
    get:
      tags: [Administration]
      summary: All accounts
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/User"
                        - type: object
                          required: [disabled, createdAt]
                          properties:
                            disabled: { type: boolean }
                            createdAt: { type: string, format: date-time }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [Administration]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { $ref: "#/components/schemas/Username" }
                password: { $ref: "#/components/schemas/Password" }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { type: object, required: [id], properties: { id: { type: string } } }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users/{id}:
    patch:
      tags: [Administration]
      summary: Disable/enable, set a password or change role
      description: Disabling or setting a password ends the user's sessions. At least one active administrator must remain.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                disabled: { type: boolean }
                password: { $ref: "#/components/schemas/Password" }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/settings/jellyfin:
    put:
      tags: [Administration]
      summary: Set the Jellyfin server URL (empty to clear)
      description: A different server forgets every user's Jellyfin token and stops all streams.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [url], properties: { url: { type: string } } }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/admin/settings/lidarr:
    put:
      tags: [Administration]
      summary: Set the Lidarr connection and acquisition profile
      description: |
        The connection is verified before saving. A blank `apiKey` keeps the saved key
        only when the URL is unchanged. Switching installations restarts every
        unfinished request there, and is refused while a Lidarr change is unconfirmed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string }
                apiKey: { type: string }
                rootFolderPath: { type: string }
                qualityProfileID: { type: integer, minimum: 0 }
                metadataProfileID: { type: integer, minimum: 0 }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/admin/lidarr/options:
    get:
      tags: [Administration]
      summary: Root folders and profiles from the saved Lidarr
      responses:
        "200":
          description: Options
          content:
            application/json:
              schema:
                type: object
                required: [roots, qualities, metadata]
                properties:
                  roots: { type: array, items: { type: object, required: [id, path], properties: { id: { type: integer }, path: { type: string } } } }
                  qualities: { type: array, items: { $ref: "#/components/schemas/Option" } }
                  metadata: { type: array, items: { $ref: "#/components/schemas/Option" } }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/acquisitions/{id}/retry:
    post:
      tags: [Administration]
      summary: Retry a failed acquisition or one needing attention
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/admin/acquisitions/{id}:
    delete:
      tags: [Administration]
      summary: Stop tracking an acquisition and remove its requests (Lidarr is not changed)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/connections/jellyfin:
    put:
      tags: [Connections]
      summary: Sign in to Jellyfin; only the resulting token is stored (encrypted)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "409": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/connections/lastfm:
    put:
      tags: [Connections]
      summary: Connect a Last.fm username and API key (read-only, public data)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, apiKey]
              properties:
                username: { type: string }
                apiKey: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/connections/{service}:
    delete:
      tags: [Connections]
      summary: Disconnect a service
      parameters:
        - { name: service, in: path, required: true, schema: { type: string, enum: [jellyfin, lastfm] } }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
...
components:
  securitySchemes:
    cookie: { type: apiKey, in: cookie, name: __Host-leerr }
    bearer: { type: http, scheme: bearer }
  responses:
    Empty:
      description: Success
      content:
        application/json:
          schema: { type: object }
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
  schemas:
    Username: { type: string, minLength: 1, maxLength: 64, pattern: "^[a-zA-Z0-9_.-]+$" }
    Password: { type: string, minLength: 10, maxLength: 256 }
    Role: { type: string, enum: [admin, member] }
    User:
      type: object
      required: [id, username, role]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    Option:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...